requirements.txt
amqp==2.5.0
atomicwrites==1.3.0
attrs==19.1.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
get==2019.4.13
idna==2.8
importlib-metadata==0.18
kombu==4.6.3
mongoengine==0.18.2
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
post==2019.4.13
public==2019.4.13
py==1.8.0
pymongo==3.8.0
pyparsing==2.4.0
pytest==4.6.3
pytest-cov==2.7.1
python-dotenv==0.10.3
pytz==2019.1
query-string==2019.4.13
requests==2.22.0
six==1.12.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
zipp==0.5.1
